😎 Feat: Add Exit Emotes (#101) 😎
* Feat: Add Exit Emotes

This would add the option to add "exit emotes" for any emote.

The point of adding this would be to allow for more "smooth" or "dynamic" animations to be set up, hopefully improving the "natural flow" when using emotes. Further down the line entry emotes could also be added to build on top of this.

This PR only includes getting up (getup) as an example on how it could be used, more emotes could be added as time goes by.

I don't think I've seen any other menu with this, so it certainly should be a step ahead of other menus.

Here is a video showing it in action: https://youtu.be/uEWAz-xgHlg (sorry for the "low" fps, my pc suddenly decided to not cooperate)

* Made Requested Changes

Add checks to confirm that both the exit type and emote exists. ExitEmoteType also defaults to "Emotes" if not specified.

ExitEmote config have been changed to match the style of other configs. It is now split up into ExitEmote and ExitEmoteType.

Added new array in the RP table called "Exits" to allow adding exit emotes without having them appear in the menu, they are still available trough commands.