Update Emote.lua (#80)
* Update Emote.lua

The code was trying to use ename which was not defined anywhere in the scope. The intended variable was actually the name parameter passed to the OnEmotePlay function.

I fix this by simply uses the correct name parameter instead of the undefined ename variable. This will properly show which emote failed to load.(If any)

* Removed unnecessary comments!