allow moving through the menu without pause
This solves the problem of moving through the emote menu and getting "stuck" when going over an emote that attaches a prop and uses emoteduration.

I haven't blocked cleaning up of props on the actual ped, because all of these props are cigarettes, cigars, joints, etc which you would want to stay on your ped in most cases and can be cleared by press X again.